Implement project maintenance improvements
