+++ pyproject.toml
@@ -32,6 +32,7 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.9"
+tomli = { version = "^2.0", python = "<3.11" }
 
 [tool.poetry.dev-dependencies]
 pytest = "^7.2.1"
@@ -43,6 +44,9 @@ testpaths = ["tests"]
 # as documented in the README.
 addopts = ""
 
+[tool.poetry-dynamic-versioning]
+enable = true
+
 [build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
+build-backend = "poetry_dynamic_versioning.backend"
