+++ rutificador/main.py
@@ -9,17 +9,24 @@ extensibility features.
 import json
 import logging
 import re
-from abc import ABC, abstractmethod
 from contextlib import contextmanager
 from dataclasses import dataclass, field
 from enum import Enum
 from functools import lru_cache, wraps
 from typing import (
-    Any, Dict, List, Literal, Optional, Protocol, Sequence, 
+    Any, Dict, List, Literal, Optional, Protocol, Sequence,
     Tuple, Union, runtime_checkable, Iterator, Type
 )
 import warnings
 from concurrent.futures import ThreadPoolExecutor, as_completed
+
+from .formatter import (
+    RutFormatter,
+    CSVFormatter,
+    XMLFormatter,
+    JSONFormatter,
+    RutFormatterFactory,
+)
 
 # ============================================================================
 # LOGGING CONFIGURATION
@@ -409,146 +416,6 @@ class RutValidator:
 # ENHANCED FORMATTING LAYER
 # ============================================================================
 
-class RutFormatter(ABC):
-    """Abstract base class for RUT formatters with validation."""
-    
-    @abstractmethod
-    def format(self, ruts: Sequence[str]) -> str:
-        """Format a sequence of RUTs."""
-        pass
-    
-    def validate_input(self, ruts: Sequence[str]) -> None:
-        """Validate input before formatting."""
-        if not isinstance(ruts, (list, tuple)):
-            raise TypeError(f"Expected sequence, got {type(ruts).__name__}")
-        
-        if not ruts:
-            logger.warning("Empty RUT sequence provided for formatting")
-
-class CSVFormatter(RutFormatter):
-    """Enhanced CSV formatter with configurable options."""
-    
-    def __init__(self, header: str = "rut", delimiter: str = "\n") -> None:
-        """
-        Initialize CSV formatter with customizable options.
-        
-        Args:
-            header: Column header name.
-            delimiter: Row delimiter.
-        """
-        self.header = header
-        self.delimiter = delimiter
-    
-    def format(self, ruts: Sequence[str]) -> str:
-        """Format RUTs as CSV with validation."""
-        self.validate_input(ruts)
-        
-        if not ruts:
-            return f"{self.header}{self.delimiter}"
-        
-        # Escape any commas in RUT strings for proper CSV format
-        escaped_ruts = [str(rut).replace(',', '\\,') for rut in ruts]
-        cadena_ruts = self.delimiter.join(escaped_ruts)
-        
-        return f"{self.header}{self.delimiter}{cadena_ruts}"
-
-class XMLFormatter(RutFormatter):
-    """Enhanced XML formatter with proper escaping and validation."""
-    
-    def __init__(self, root_element: str = "root", item_element: str = "rut") -> None:
-        """
-        Initialize XML formatter with customizable element names.
-        
-        Args:
-            root_element: Name of the root XML element.
-            item_element: Name of individual RUT elements.
-        """
-        self.root_element = root_element
-        self.item_element = item_element
-    
-    def format(self, ruts: Sequence[str]) -> str:
-        """Format RUTs as XML with proper escaping."""
-        self.validate_input(ruts)
-        
-        xml_lines = [f"<{self.root_element}>"]
-        
-        for rut in ruts:
-            # Comprehensive XML escaping
-            rut_escaped = (str(rut)
-                          .replace("&", "&amp;")
-                          .replace("<", "&lt;")
-                          .replace(">", "&gt;")
-                          .replace('"', "&quot;")
-                          .replace("'", "&#x27;"))
-            xml_lines.append(f"    <{self.item_element}>{rut_escaped}</{self.item_element}>")
-        
-        xml_lines.append(f"</{self.root_element}>")
-        return "\n".join(xml_lines)
-
-class JSONFormatter(RutFormatter):
-    """Enhanced JSON formatter with configurable structure."""
-    
-    def __init__(self, key_name: str = "rut", pretty_print: bool = True) -> None:
-        """
-        Initialize JSON formatter with customizable options.
-        
-        Args:
-            key_name: Key name for RUT values in JSON objects.
-            pretty_print: Whether to format JSON with indentation.
-        """
-        self.key_name = key_name
-        self.pretty_print = pretty_print
-    
-    def format(self, ruts: Sequence[str]) -> str:
-        """Format RUTs as JSON with validation."""
-        self.validate_input(ruts)
-        
-        ruts_json = [{self.key_name: str(rut)} for rut in ruts]
-        
-        return json.dumps(
-            ruts_json, 
-            indent=2 if self.pretty_print else None,
-            ensure_ascii=False,
-            separators=(',', ': ') if self.pretty_print else (',', ':')
-        )
-
-class RutFormatterFactory:
-    """
-    Enhanced factory for creating RUT formatters with configuration support.
-    
-    This factory implements the Factory pattern with support for formatter
-    configuration and extensibility through registration of custom formatters.
-    """
-    
-    _formatters: Dict[str, Type[RutFormatter]] = {
-        "csv": CSVFormatter,
-        "xml": XMLFormatter,
-        "json": JSONFormatter,
-    }
-    
-    @classmethod
-    def register_formatter(cls, name: str, formatter_class: Type[RutFormatter]) -> None:
-        """Register a custom formatter."""
-        if not issubclass(formatter_class, RutFormatter):
-            raise TypeError("Formatter must inherit from RutFormatter")
-        cls._formatters[name.lower()] = formatter_class
-        logger.info(f"Registered custom formatter: {name}")
-    
-    @classmethod
-    def get_formatter(cls, formato: str, **kwargs: Any) -> Optional[RutFormatter]:
-        """Get formatter instance with optional configuration."""
-        if not isinstance(formato, str):
-            return None
-        
-        formatter_class = cls._formatters.get(formato.lower())
-        if formatter_class:
-            return formatter_class(**kwargs)
-        return None
-    
-    @classmethod
-    def get_available_formats(cls) -> List[str]:
-        """Get list of available formatter names."""
-        return list(cls._formatters.keys())
 
 # ============================================================================
 # ENHANCED DOMAIN MODELS
@@ -1159,8 +1026,10 @@ def get_version_info() -> Dict[str, str]:
     Returns:
         Dictionary with version and module metadata.
     """
+    from . import __version__
+
     return {
-        "version": "0.4.1",
+        "version": __version__,
         "author": "Carlos Ortega González",
         "license": "MIT",
         "description": "Enhanced Chilean RUT validation and formatting library",
@@ -1227,9 +1096,6 @@ def benchmark_performance(num_ruts: int = 10000,
 # ============================================================================
 # MODULE INITIALIZATION
 # ============================================================================
-
-# Set up default logging
-setup_logging()
 
 # Log module initialization
 logger.info("Enhanced Rutificador module initialized successfully")
